refactor(devframe)!: split auth into `devframe/node/auth`, move `devframe/internal` to `devframe/node/internal`
BREAKING CHANGE: auth helpers (`getTempAuthToken`, `refreshTempAuthToken`,
`consumeTempAuthToken`, `getPendingAuth`, `setPendingAuth`, `abortPendingAuth`,
`PendingAuthRequest`, `revokeAuthToken`, `revokeActiveConnectionsForToken`)
no longer re-export from `devframe/node` — import from `devframe/node/auth`.

The `devframe/internal` sub-export is renamed to `devframe/node/internal`,
which also now hosts the relocated `context-internal.ts`. End users were
never meant to depend on this subpath; first-party adapters update accordingly.
