+++ alias.ts
@@ -13,9 +13,10 @@ export const alias = {
   'devframe/rpc/server': df('devframe/src/rpc/server.ts'),
   'devframe/rpc': df('devframe/src/rpc'),
   'devframe/types': df('devframe/src/types/index.ts'),
+  'devframe/node/auth': df('devframe/src/node/auth/index.ts'),
+  'devframe/node/internal': df('devframe/src/node/internal/index.ts'),
   'devframe/node': df('devframe/src/node/index.ts'),
   'devframe/constants': df('devframe/src/constants.ts'),
-  'devframe/internal': df('devframe/src/internal/index.ts'),
   'devframe/utils/events': df('devframe/src/utils/events.ts'),
   'devframe/utils/human-id': df('devframe/src/utils/human-id.ts'),
   'devframe/utils/nanoid': df('devframe/src/utils/nanoid.ts'),
