+++ devframe/packages/devframe/package.json
@@ -28,8 +28,9 @@
     "./adapters/vite": "./dist/adapters/vite.mjs",
     "./client": "./dist/client/index.mjs",
     "./constants": "./dist/constants.mjs",
-    "./internal": "./dist/internal/index.mjs",
     "./node": "./dist/node/index.mjs",
+    "./node/auth": "./dist/node/auth.mjs",
+    "./node/internal": "./dist/node/internal.mjs",
     "./recipes/open-helpers": "./dist/recipes/open-helpers.mjs",
     "./rpc": "./dist/rpc/index.mjs",
     "./rpc/client": "./dist/rpc/client.mjs",
