+++ devframe/packages/devframe/tsdown.config.ts
@@ -10,8 +10,9 @@ export default defineConfig({
     'rpc/transports/ws-server': 'src/rpc/transports/ws-server.ts',
     'types/index': 'src/types/index.ts',
     'node/index': 'src/node/index.ts',
+    'node/auth': 'src/node/auth/index.ts',
+    'node/internal': 'src/node/internal/index.ts',
     'constants': 'src/constants.ts',
-    'internal/index': 'src/internal/index.ts',
     'utils/events': 'src/utils/events.ts',
     'utils/human-id': 'src/utils/human-id.ts',
     'utils/nanoid': 'src/utils/nanoid.ts',
