+++ packages/core/src/index.ts
@@ -2,4 +2,4 @@ export { createDevToolsContext } from './node/context'
 export { DevTools } from './node/plugins'
 export type { BuiltinServerFunctions } from './node/rpc'
 export { createDevToolsMiddleware } from './node/server'
-export type { DevToolsInternalContext, InternalAnonymousAuthStorage } from 'devframe/node'
+export type { DevToolsInternalContext, InternalAnonymousAuthStorage } from 'devframe/node/internal'
