+++ packages/core/src/internal.ts
@@ -1,2 +1,2 @@
-export { getInternalContext } from 'devframe/node'
-export type { DevToolsInternalContext, InternalAnonymousAuthStorage } from 'devframe/node'
+export { getInternalContext } from 'devframe/node/internal'
+export type { DevToolsInternalContext, InternalAnonymousAuthStorage } from 'devframe/node/internal'
