+++ packages/core/src/node/rpc/anonymous/auth.ts
@@ -3,7 +3,8 @@ import process from 'node:process'
 import * as p from '@clack/prompts'
 import { defineRpcFunction } from '@vitejs/devtools-kit'
 import c from 'ansis'
-import { abortPendingAuth, getInternalContext, getTempAuthToken, refreshTempAuthToken, setPendingAuth } from 'devframe/node'
+import { abortPendingAuth, getTempAuthToken, refreshTempAuthToken, setPendingAuth } from 'devframe/node/auth'
+import { getInternalContext } from 'devframe/node/internal'
 import { MARK_INFO } from '../../constants'
 
 export interface DevToolsAuthInput {
