+++ packages/core/src/node/ws.ts
@@ -5,7 +5,7 @@ import type { WebSocket } from 'ws'
 import { AsyncLocalStorage } from 'node:async_hooks'
 import process from 'node:process'
 import c from 'ansis'
-import { getInternalContext } from 'devframe/node'
+import { getInternalContext } from 'devframe/node/internal'
 import { createRpcServer } from 'devframe/rpc/server'
 import { attachWsRpcTransport } from 'devframe/rpc/transports/ws-server'
 import { getPort } from 'get-port-please'
