+++ packages/kit/src/node/host-docks.ts
@@ -12,8 +12,8 @@ import type {
 import type { DevToolsDocksUserSettings } from '../types/settings'
 import type { KitNodeContext } from './context'
 import { REMOTE_CONNECTION_KEY } from 'devframe/constants'
-import { getInternalContext } from 'devframe/internal'
 import { createStorage } from 'devframe/node'
+import { getInternalContext } from 'devframe/node/internal'
 import { createEventEmitter } from 'devframe/utils/events'
 import { join } from 'pathe'
 import { DEFAULT_STATE_USER_SETTINGS } from '../constants'
